Refactor IntCodeComputer and Utils

+++ src/test/kotlin/Day5Spec.kt
@@ -9,14 +9,14 @@ object Day5Spec: Spek({
     given("AoC 19.5.1") {
 
         it("works with task data") {
-            assertEquals(6745903, Day5.solvePt1(readLines("5.txt").first().split(",").map { it.toInt() }))
+            assertEquals(6745903, Day5.solvePt1(readLines("5.txt").first().split(",").map { it.toLong() }))
         }
     }
 
     given("AoC 19.5.2") {
 
         it("works with task data") {
-            assertEquals(9168267, Day5.solvePt2(readLines("5.txt").first().split(",").map { it.toInt() }))
+            assertEquals(9168267, Day5.solvePt2(readLines("5.txt").first().split(",").map { it.toLong() }))
         }
 
     }
